Archive pre-2019 Heroku site; Update README.md
Excerpt from the  new README:

```
## timeline
2015: first bingo! [eli](https://twitter.com/veryeli) and i used the attendee
directory to generate a unique card for everyone (twitter login kept it private
🔒). squares on your card were other attendees - if you met someone on your
card you got to check it off. we made it cuz we’re shy. most of it is in the
`pre-2019` folder!

2016: we made the cards prettier by pulling in people’s twitter photos and
doing imgmagick to them 🔮

2017: no xoxo, no bingo… missed u all

2018: xoxo was in the midst of changing their infrastructure, so i lost access
to the attendee directory. [hannah](https://twitter.com/herlifeinpixels),
[jason](https://twitter.com/justsomeguy) and i met in a cafe before the kickoff
ceremony and designed a static version with input from the community. hannah
and jason made 25 icons in like two minutes, it was incredible!!!

2019: i've been too cheap to get https://xoxo.bingo in previous years, but
[andy](https://twitter.com/andymcmillan) noticed a thread on slack and hooked
us up! thx andy.
```

leading up to xoxo2018, i realized we wouldn't have access to the new
attendee registry. [andy](https://twitter.com/waxpancake) and i
discussed ad-hoc private access and other ways to make it work, but it
was too much. so hannah, jason and i made a static version with
"achievements" sourced from the slack community.

it was fun to get excited about things specific to that year, like the
podcast airstream and the blue ox. and it's gonna be that way from now
on! feel free to create an issue or msg on slack if you have ideas for
this year's bingo squares.

since it's staying a static site, i moved everything off of heroku.
all site content will live in [`2019-and-on/`](./2019-and-on/).